[OTS-680] Co-browsing does not draw under a circumstance. (#127)
By German suggestion: We detected that this bug occurs “randomly”. Sometimes web doesn’t receive the  signal. Theoretically you need to remove the connection on this line
We think that  sometimes is wrong, so web is not receiving the signal. We replace that for  and it started to work on all the tests.

Tested with Andrea y German. Andrea: its probably gonna be useful for multiparty as well because if you have a connection then its gonna restrict the signaling to it, you need to send it to all in the session